realize: rewrite suspicious append with assignment
The append in form of `a = append(b, ...)` is suspicious
due to mistatching `a` and `b`.

In this particular case, there is a subtle syntax-level
difference due to embedding, but same field is being assigned,
so more "safe" looking code is something like:

	r.Projects = append(r.Projects, Project{exit: make(chan os.Signal, 1)})

But even better, this can be re-written as assignment, without append.